+++ HISTORY.rst
@@ -6,6 +6,14 @@ History
 
 - Bumped asyncio-rpc to 0.1.10
 
+- Fixed GeoJSON export with pyproj <= 1.9.6.
+
+- Fixed compatibility with h5py 3.*.
+
+- Added gridadmin.cells.transform.
+
+- Fixed gridadmin.grid.n2dtot and .dx propagation.
+
 
 1.0.27 (2021-02-22)
 -------------------